feat: resolve unzip using package.json browser field
